+++ src/main/java/in/koreatech/koin/admin/shop/repository/menu/AdminMenuRepository.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.admin.shop.repository;
+package in.koreatech.koin.admin.shop.repository.menu;
 
 import java.util.List;
 import java.util.Optional;
