+++ src/main/java/in/koreatech/koin/admin/shop/repository/review/AdminShopReviewRepository.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.admin.shop.repository;
+package in.koreatech.koin.admin.shop.repository.review;
 
 import java.util.Optional;
